+++ packages/docs/src/stories/Box.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { Box, BoxProps, Text, Heading } from "@ignite-ui/react";
+import { Box, BoxProps, Text, Heading } from "@pedrogiampietro-ignite-ui/react";
 
 export default {
   title: "Surfaces/Box",