+++ packages/react/src/styles/index.ts
@@ -7,7 +7,7 @@ import {
   lineHeights,
   radii,
   space,
-} from "@ignite-ui/tokens";
+} from "@pedrogiampietro-ignite-ui/tokens";
 import { createStitches, defaultThemeMap } from "@stitches/react";
 
 export const {